minor cleanup + starting rendering system

+++ include/Logic.hpp
@@ -20,7 +20,7 @@ private:
 
   bool gameOver;
 
-  Vect<2u, float> mousePos;
+  Vect<2u, double> mousePos;
 
   void handleKey(GLFWwindow *window, Key key);
   void handleMouse(Display const &, GLFWwindow *window, Mouse mouse);
@@ -37,7 +37,7 @@ public:
   void incCombo();
 
   Vect<2, double> getPlayerPos(void) const;
-  Vect<2u, float> getMouse(Display const &) const;
+  Vect<2u, double> getMouse(Display const &) const;
   unsigned int  getScore(void) const;
   std::string   getTime(void) const;
   std::string   getCombo(void) const;
